added FriendsForm component, styling
