+++ friends/src/App.js
@@ -5,18 +5,19 @@ import Login from './components/Login';
 import FriendsList from './components/FriendsList';
 import FriendsCard from './components/FriendsCard';
 import PrivateRoute from './components/PrivateRoute';
+import header from './images/header.png';
 import './App.css';
 
 const App = () => {
   return (
     <Router >
       <div className="App">
+        <div className='empty-container'></div>
+        <img className='header' src={header} />
         <nav>
           <Link className='nav-link' to="/">Home</Link>
           <Link className='nav-link' to="/api/login">Login</Link>
-          <Link className='nav-link' to='/api/friends'>Friends List</Link>
         </nav>
-        <h1>Friends</h1>
       </div>
       <Switch>
         <Route exact path='/api/login' component={Login} />
